+++ next.config.ts
@@ -1,5 +1,16 @@
 import type { NextConfig } from 'next';
 
-const nextConfig: NextConfig = {};
+const nextConfig: NextConfig = {
+  agentRules: false,
+  output: 'export',
+  trailingSlash: true,
+  poweredByHeader: false,
+  images: {
+    unoptimized: true,
+  },
+  turbopack: {
+    root: process.cwd(),
+  },
+};
 
 export default nextConfig;